refactor: enhance code readability and structure in SpeechTranscriber and related files
