+++ Sources/SpeechTranscriber.swift
@@ -191,7 +191,8 @@ class SpeechTranscriber {
             let progress = Double(currentSegmentCount) / Double(estimatedTotalSegments)
             let progressPercent = String(format: "%.1f%%", progress * 100)
             let audioProcessedStr = String(format: "%.1f", audioProcessed)
-            let progressMessage = "DEBUG: Transcription progress: \(progressPercent) - \(currentSegmentCount) segments, \(audioProcessedStr)s processed\n"
+            let progressMessage = "DEBUG: Transcription progress: \(progressPercent) - \(currentSegmentCount) segments, " +
+                "\(audioProcessedStr)s processed\n"
             fputs(progressMessage, stderr)
             // TEMP DEBUG: Bypass Logger call
             // let progressPercent = String(format: "%.1f%%", progress * 100)
@@ -343,110 +344,6 @@ class SpeechTranscriber {
     }
 }
 
-// MARK: - Language Detection Extension
-
-extension SpeechTranscriber {
-    /// Validate and normalize language code
-    static func validateLanguageCode(_ languageCode: String) -> String? {
-        fputs("DEBUG: In validateLanguageCode for: \(languageCode)\n", stderr)
-
-        // TEMP DEBUG: Bypass supportedLocales() call which might cause hang
-        fputs("DEBUG: Bypassing supportedLocales() call for debugging\n", stderr)
-        // let supportedLocales = Self.supportedLocales()
-        // let supportedIdentifiers = Set(supportedLocales.map { $0.identifier })
-
-        // Simple fallback for debugging - just return the input if it looks like a valid language code
-        if languageCode == "en-US" || languageCode == "en" {
-            fputs("DEBUG: Validated \(languageCode) as en-US\n", stderr)
-            return "en-US"
-        }
-
-        fputs("DEBUG: Language code '\(languageCode)' not in simple validation, returning en-US fallback\n", stderr)
-        return "en-US" // Simple fallback for debugging
-
-        /*
-         // Original implementation commented out for debugging
-         // Try exact match first
-         if supportedIdentifiers.contains(languageCode) {
-         return languageCode
-         }
-
-         // Try language-only match (e.g., "en" -> "en-US")
-         let languageOnly = String(languageCode.prefix(2))
-         if let match = supportedIdentifiers.first(where: { $0.hasPrefix(languageOnly + "-") }) {
-         // TEMP DEBUG: Bypass Logger call
-         // Logger.shared.info("Language code '\(languageCode)' normalized to '\(match)'", component: "SpeechTranscriber")
-         return match
-         }
-
-         // Try case-insensitive match
-         if let match = supportedIdentifiers.first(where: { $0.lowercased() == languageCode.lowercased() }) {
-         // TEMP DEBUG: Bypass Logger call
-         // Logger.shared.info("Language code '\(languageCode)' normalized to '\(match)'", component: "SpeechTranscriber")
-         return match
-         }
-
-         // TEMP DEBUG: Bypass Logger call
-         // Logger.shared.warn("Language code '\(languageCode)' is not supported", component: "SpeechTranscriber")
-         return nil
-         */
-    }
-
-    /// Get list of common language codes for easier user reference
-    static func commonLanguages() -> [String: String] {
-        return [
-            "en-US": "English (United States)",
-            "en-GB": "English (United Kingdom)",
-            "es-ES": "Spanish (Spain)",
-            "es-MX": "Spanish (Mexico)",
-            "fr-FR": "French (France)",
-            "de-DE": "German (Germany)",
-            "it-IT": "Italian (Italy)",
-            "pt-BR": "Portuguese (Brazil)",
-            "ja-JP": "Japanese (Japan)",
-            "ko-KR": "Korean (South Korea)",
-            "zh-CN": "Chinese (Simplified)",
-            "zh-TW": "Chinese (Traditional)",
-            "ru-RU": "Russian (Russia)",
-            "ar-SA": "Arabic (Saudi Arabia)"
-        ]
-    }
-
-    /// Transcribe with automatic language detection and validation
-    func transcribeWithLanguageDetection(
-        audioFile: AudioFile,
-        preferredLanguages: [String] = ["en-US"],
-        progressCallback: ProgressCallback? = nil
-    ) async throws -> TranscriptionResult {
-        fputs("DEBUG: In transcribeWithLanguageDetection\n", stderr)
-        // Validate and normalize all preferred languages
-        fputs("DEBUG: About to validate languages\n", stderr)
-        let validatedLanguages = preferredLanguages.compactMap { Self.validateLanguageCode($0) }
-        fputs("DEBUG: Languages validated\n", stderr)
-
-        if validatedLanguages.isEmpty {
-            fputs("DEBUG: No valid languages provided, falling back to en-US\n", stderr)
-            // TEMP DEBUG: Bypass Logger call
-            // Logger.shared.warn("No valid languages provided, falling back to en-US", component: "SpeechTranscriber")
-            let fallbackLanguages = ["en-US"]
-            return try await attemptTranscriptionWithLanguages(
-                audioFile: audioFile,
-                languages: fallbackLanguages,
-                progressCallback: progressCallback
-            )
-        }
-
-        fputs("DEBUG: Validated languages: \(validatedLanguages.joined(separator: ", "))\n", stderr)
-        // TEMP DEBUG: Bypass Logger call
-        // Logger.shared.info("Validated languages: \(validatedLanguages.joined(separator: ", "))", component: "SpeechTranscriber")
-        fputs("DEBUG: About to call attemptTranscriptionWithLanguages\n", stderr)
-        return try await attemptTranscriptionWithLanguages(
-            audioFile: audioFile,
-            languages: validatedLanguages,
-            progressCallback: progressCallback
-        )
-    }
-}
 
 // MARK: - Private Language Processing Extension
 
@@ -468,64 +365,26 @@ private extension SpeechTranscriber {
         // Try each language in order
         fputs("DEBUG: About to iterate languages: \(languages.joined(separator: ", "))\n", stderr)
         for (index, languageCode) in languages.enumerated() {
-            fputs("DEBUG: Trying language \(languageCode) (\(index + 1)/\(languages.count))\n", stderr)
-            let locale = Locale(identifier: languageCode)
-
-            fputs("DEBUG: About to check if speech recognition is available for \(languageCode)\n", stderr)
-            guard Self.isAvailable(for: locale) else {
-                fputs("DEBUG: Speech recognition not available for \(languageCode)\n", stderr)
-                // TEMP DEBUG: Bypass Logger call
-                // Logger.shared.warn("Speech recognition not available for \(languageCode)", component: "SpeechTranscriber")
-                continue
-            }
-
-            let attemptMessage = "DEBUG: Attempting transcription with language: \(languageCode) (\(index + 1)/\(languages.count))\n"
-            fputs(attemptMessage, stderr)
-            // TEMP DEBUG: Bypass Logger call
-            // Logger.shared.info("Attempting transcription with language: \(languageCode) (\(index + 1)/\(languages.count))", component: "SpeechTranscriber")
-
             do {
-                fputs("DEBUG: About to create SpeechTranscriber with locale \(languageCode)\n", stderr)
-                let speechTranscriber = try SpeechTranscriber(locale: locale)
-                fputs("DEBUG: SpeechTranscriber created successfully, about to call transcribe\n", stderr)
-                let result = try await speechTranscriber.transcribe(
+                if let result = try await attemptTranscriptionWithLanguage(
+                    languageCode: languageCode,
+                    index: index,
+                    totalCount: languages.count,
                     audioFile: audioFile,
-                    progressCallback: progressCallback
-                )
-                fputs("DEBUG: transcribe() completed successfully\n", stderr)
-
-                Logger.shared.info(
-                    "Transcription confidence for \(languageCode): \(String(format: "%.1f%%", result.confidence * 100))",
-                    component: "SpeechTranscriber"
-                )
-
-                // Use ConfidenceManager to assess quality
-                let qualityAssessment = confidenceManager.assessQuality(result: result)
-
-                // Log quality assessment
-                Logger.shared.info(
-                    "Quality level: \(qualityAssessment.qualityLevel.rawValue)",
-                    component: "SpeechTranscriber"
-                )
-                if !qualityAssessment.lowConfidenceSegments.isEmpty {
-                    Logger.shared.warn(
-                        "Found \(qualityAssessment.lowConfidenceSegments.count) low-confidence segments",
-                        component: "SpeechTranscriber"
-                    )
-                }
-
-                // If confidence meets quality standards, return immediately
-                if confidenceManager.meetsQualityStandards(result: result) {
-                    Logger.shared.info(
-                        "High quality result achieved with \(languageCode)",
-                        component: "SpeechTranscriber"
-                    )
-                    return result
-                }
-
-                // Keep the best result so far
-                if bestResult == nil || result.confidence > (bestResult?.confidence ?? 0.0) {
-                    bestResult = result
+                    progressCallback: progressCallback,
+                    confidenceManager: confidenceManager
+                ) {
+                    if confidenceManager.meetsQualityStandards(result: result) {
+                        Logger.shared.info(
+                            "High quality result achieved with \(languageCode)",
+                            component: "SpeechTranscriber"
+                        )
+                        return result
+                    }
+                    
+                    if bestResult == nil || result.confidence > (bestResult?.confidence ?? 0.0) {
+                        bestResult = result
+                    }
                 }
             } catch {
                 Logger.shared.warn(
@@ -537,16 +396,75 @@ private extension SpeechTranscriber {
             }
         }
 
-        // Return the best result we got, with quality assessment warnings
+        return try processBestResult(bestResult: bestResult, confidenceManager: confidenceManager, lastError: lastError)
+    }
+    
+    private func attemptTranscriptionWithLanguage(
+        languageCode: String,
+        index: Int,
+        totalCount: Int,
+        audioFile: AudioFile,
+        progressCallback: ProgressCallback?,
+        confidenceManager: ConfidenceManager
+    ) async throws -> TranscriptionResult? {
+        fputs("DEBUG: Trying language \(languageCode) (\(index + 1)/\(totalCount))\n", stderr)
+        let locale = Locale(identifier: languageCode)
+
+        fputs("DEBUG: About to check if speech recognition is available for \(languageCode)\n", stderr)
+        guard Self.isAvailable(for: locale) else {
+            fputs("DEBUG: Speech recognition not available for \(languageCode)\n", stderr)
+            return nil
+        }
+
+        let attemptMessage = "DEBUG: Attempting transcription with language: \(languageCode) (\(index + 1)/\(totalCount))\n"
+        fputs(attemptMessage, stderr)
+
+        fputs("DEBUG: About to create SpeechTranscriber with locale \(languageCode)\n", stderr)
+        let speechTranscriber = try SpeechTranscriber(locale: locale)
+        fputs("DEBUG: SpeechTranscriber created successfully, about to call transcribe\n", stderr)
+        let result = try await speechTranscriber.transcribe(
+            audioFile: audioFile,
+            progressCallback: progressCallback
+        )
+        fputs("DEBUG: transcribe() completed successfully\n", stderr)
+
+        Logger.shared.info(
+            "Transcription confidence for \(languageCode): \(String(format: "%.1f%%", result.confidence * 100))",
+            component: "SpeechTranscriber"
+        )
+
+        return logQualityAssessment(result: result, confidenceManager: confidenceManager)
+    }
+    
+    private func logQualityAssessment(result: TranscriptionResult, confidenceManager: ConfidenceManager) -> TranscriptionResult {
+        let qualityAssessment = confidenceManager.assessQuality(result: result)
+
+        Logger.shared.info(
+            "Quality level: \(qualityAssessment.qualityLevel.rawValue)",
+            component: "SpeechTranscriber"
+        )
+        if !qualityAssessment.lowConfidenceSegments.isEmpty {
+            Logger.shared.warn(
+                "Found \(qualityAssessment.lowConfidenceSegments.count) low-confidence segments",
+                component: "SpeechTranscriber"
+            )
+        }
+        
+        return result
+    }
+    
+    private func processBestResult(
+        bestResult: TranscriptionResult?,
+        confidenceManager: ConfidenceManager,
+        lastError: Error?
+    ) throws -> TranscriptionResult {
         if let result = bestResult {
             let qualityAssessment = confidenceManager.assessQuality(result: result)
 
-            // Log quality warnings
             for warning in qualityAssessment.warnings {
                 Logger.shared.warn(warning, component: "SpeechTranscriber")
             }
 
-            // Log fallback recommendation
             if qualityAssessment.shouldUseFallback {
                 Logger.shared.warn(
                     "Transcription quality is low - consider using cloud fallback services",
@@ -556,7 +474,7 @@ private extension SpeechTranscriber {
 
             return result
         }
-        // If we got here, all languages failed
+        
         if let error = lastError {
             throw error
         } else {
@@ -565,77 +483,3 @@ private extension SpeechTranscriber {
     }
 }
 
-// MARK: - Segment Processing Extension
-
-private extension SpeechTranscriber {
-    func determineSegmentType(
-        text: String,
-        pauseDuration: TimeInterval?,
-        isFirstSegment: Bool,
-        isLastSegment: Bool,
-        previousText: String?
-    ) -> SegmentType {
-        // Check for silence gaps (pause > speaker change threshold typically indicates speaker change or paragraph break)
-        if let pause = pauseDuration, pause > TimingThresholds.speakerChangeThreshold {
-            return .speakerChange
-        }
-
-        // Check for sentence boundaries
-        if text.hasSuffix(".") || text.hasSuffix("!") || text.hasSuffix("?") {
-            return .sentenceBoundary
-        }
-
-        // Check for paragraph boundaries (long pause + sentence ending)
-        if let pause = pauseDuration, pause > TimingThresholds.paragraphBoundaryThreshold,
-           let prevText = previousText,
-           prevText.hasSuffix(".") || prevText.hasSuffix("!") || prevText.hasSuffix("?") {
-            return .paragraphBoundary
-        }
-
-        // Check for silence (empty or very short segments with low confidence)
-        if text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            return .silence
-        }
-
-        return .speech
-    }
-
-    func extractWordTimings(from segment: SFTranscriptionSegment) -> WordTiming? {
-        // Apple's SFTranscriptionSegment represents word-level segments
-        // Each segment typically contains one word with timing information
-        let word = segment.substring.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        guard !word.isEmpty else { return nil }
-
-        // TEMP DEBUG: Bypass Logger calls
-        if word.split(separator: " ").count > 1 {
-            fputs("DEBUG: Unexpected multi-word segment found: '\(word)'\n", stderr)
-        }
-
-        return WordTiming(
-            word: word,
-            startTime: segment.timestamp,
-            endTime: segment.timestamp + segment.duration,
-            confidence: Double(segment.confidence)
-        )
-    }
-
-    func detectSpeakerChange(at index: Int, in segments: [SFTranscriptionSegment]) -> String? {
-        // Apple Speech framework doesn't provide speaker diarization
-        // This is a placeholder for future enhancement or integration with other services
-        // For now, we'll detect potential speaker changes based on significant pause patterns
-
-        guard index > 0 else { return "Speaker1" }
-
-        let currentSegment = segments[index]
-        let previousSegment = segments[index - 1]
-        let pauseDuration = currentSegment.timestamp - (previousSegment.timestamp + previousSegment.duration)
-
-        // Simple heuristic: if there's a pause > definite speaker change threshold, assume speaker change
-        if pauseDuration > TimingThresholds.definiteSpeakerChangeThreshold {
-            return "Speaker\((index % 4) + 1)" // Cycle through up to 4 speakers
-        }
-
-        return "Speaker1" // Default to single speaker
-    }
-}
